refactor: use function for creating HistoryItemInfo objects
